builtin-types (#11)

+++ docs/examples/tutorial/profiling_tutorial/calc_pi_4.pyx
@@ -2,16 +2,13 @@
 
 cimport cython
 
-
-
-
 @cython.profile(False)
-cdef inline double recip_square(long long i) except -1.0:
+cdef inline f64 recip_square(i128 i) except -1.0:
     return 1. / (i * i)
 
-def approx_pi(int n=10000000):
-    cdef double val = 0.
-    cdef int k
+def approx_pi(i32 n=10000000):
+    cdef f64 val = 0.
+    cdef i32 k
     for k in range(1, n + 1):
         val += recip_square(k)
     return (6 * val) ** .5
